v1.2
This release of the Mason CLI fixes an APK registration bug and
introduces a new beta feature for fast Mason Config image building.

## New Features

### Turbo Builder

- Mason Config images can now be built using a new method that creates the
images significantly faster. For example, run `mason stage --turbo
config.yml`. This feature is now in beta and should be tested in a
non-production environment before deploying to critical devices.

## Improvements
- A bug was introduced in v1.1 that caused some APKs to not be accepted
for registration based on their signatures. This release fixes this
regression.