Fix cyclic imports
This one is an interesting bug because it doesn't happen in local testing, only in the binaries. The problems lies in Python's module cache: we import Config from cli.config in the root mason.py. Then, in something like the CliInitCommand, we accidentally imported Config from cli.mason. This works fine locally because that module will have been cached and can be retrieved without having to finish the mason.py module initialization. For whatever reason, this is not the case in the binary.

Signed-off-by: Alex Saveau <alexsaveau@bymason.com>