Simply configuring endpoints environment
Currently, we need to add a URL for each endpoint we wish to use. This means employees constantly have to update their endpoints.yml file with the new endpoints. The new mechanism provides both the flexibility of testing a single changed endpoint and the simplicity of only having to configure one base URL for switching environments.

Signed-off-by: Alex Saveau <alexsaveau@bymason.com>